@@ -13,43 +13,42 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.1.1",
-    "@angular/cdk": "~13.1.1",
-    "@angular/common": "~13.1.1",
-    "@angular/compiler": "~13.1.1",
-    "@angular/core": "~13.1.1",
-    "@angular/forms": "~13.1.1",
-    "@angular/material": "~13.1.1",
-    "@angular/platform-browser": "~13.1.1",
-    "@angular/platform-browser-dynamic": "~13.1.1",
-    "@angular/router": "~13.1.1",
+    "@angular/animations": "^14.1.3",
+    "@angular/cdk": "^14.1.3",
+    "@angular/common": "^14.1.3",
+    "@angular/compiler": "^14.1.3",
+    "@angular/core": "^14.1.3",
+    "@angular/forms": "^14.1.3",
+    "@angular/material": "^14.1.3",
+    "@angular/platform-browser": "^14.1.3",
+    "@angular/platform-browser-dynamic": "^14.1.3",
+    "@angular/router": "^14.1.3",
     "bootstrap": "^5.1.3",
     "bootstrap-icons": "^1.6.1",
     "font-awesome": "^4.7.0",
     "moment": "^2.29.4",
     "ng-lazyload-image": "^9.1.0",
-    "rxjs": "~6.6.3",
-    "tslib": "^2.0.0",
+    "rxjs": "~7.5.0",
+    "tslib": "^2.3.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.1.2",
-    "@angular/cli": "~13.1.2",
-    "@angular/compiler-cli": "~13.1.1",
-    "@angular/language-service": "~13.1.1",
+    "@angular-devkit/build-angular": "^14.1.3",
+    "@angular/cli": "^14.1.3",
+    "@angular/compiler-cli": "^14.1.3",
+    "@angular/language-service": "^14.1.3",
     "@types/jasmine": "~3.10.0",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
+    "@types/node": "~12.11.1",
     "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.10.0",
+    "jasmine-core": "~4.0.0",
     "jasmine-spec-reporter": "~7.0.0",
     "karma": "~6.3.16",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
+    "karma-jasmine-html-reporter": "~1.7.0",
     "ts-node": "~10.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.5.2"
+    "typescript": "~4.7.4"
   }
 }
